moving queued_item_store
