+++ packages/logging_cloudwatch/aws_logging_cloudwatch/pubspec.yaml
@@ -10,10 +10,13 @@ environment:
 
 dependencies:
   amplify_core: ">=1.4.0 <1.5.0"
+  amplify_db_common_dart: ">=0.3.0+1 <0.4.0"
   aws_common: ">=0.6.0 <0.7.0"
   aws_signature_v4: ">=0.5.0 <0.6.0"
   built_collection: ^5.1.1
   built_value: ">=8.6.0 <8.7.0"
+  collection: ^1.15.0
+  drift: ">=2.11.0 <2.12.0"
   fixnum: ^1.1.0
   intl: ">=0.18.0 <1.0.0"
   json_annotation: ">=4.8.1 <4.9.0"
